requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.3
beautifulsoup4>=4.12.3
requests>=2.31.0
pydantic>=2.7.4
python-dotenv>=1.0.1
pandas>=2.2.0
tqdm>=4.66.1 